+++ src/main/java/com/api/TaveShot/domain/recommend/dto/RecProRequestDto.java
@@ -1,13 +1,11 @@
 package com.api.TaveShot.domain.recommend.dto;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @Builder
 public class RecProRequestDto {
 
-    private int solvedRecentId;
+    private Long solvedRecentId;
 }
